Implemented the Update method and tests for Room.

+++ BookingService/Core/Application/Room/RoomManager.cs
@@ -4,7 +4,9 @@ using Application.Room.Requests;
 using Application.Room.Responses;
 using Domain.Entities;
 using Domain.Ports;
+using Domain.Room.Enums;
 using Domain.Room.Exceptions;
+using Domain.Room.ValueObjects;
 
 namespace Application.Room
 {
@@ -28,7 +30,7 @@ namespace Application.Room
                 {
                     RoomData = request.RoomData,
                     Success = true,
-                    Message = "Sala criada com sucesso!"
+                    Message = "The room has been created successfully!"
                 };
             }
             catch (InvalidRoomNameException)
@@ -37,7 +39,7 @@ namespace Application.Room
                 {
                     Success = false,
                     ErrorCode = ErrorCode.MISSING_REQUIRED_INFORMATION,
-                    Message = "Missing passed name information"
+                    Message = "Missing name information"
                 };
             }
             catch (InvalidRoomPriceException)
@@ -55,7 +57,7 @@ namespace Application.Room
                 {
                     Success = false,
                     ErrorCode = ErrorCode.MISSING_REQUIRED_INFORMATION,
-                    Message = "Missing passed level information"
+                    Message = "Missing level information"
                 };
             }
             catch (RoomNotAvailableException)
@@ -88,16 +90,36 @@ namespace Application.Room
                     return new RoomResponse
                     {
                         Success = false,
-                        Message = "Sala não encontrada."
+                        ErrorCode = ErrorCode.NOT_FOUND,
+                        Message = "Room not found."
                     };
                 }
 
-                await _roomRepository.Delete(roomId);
+                if (room.Bookings != null && room.Bookings.Any())
+                {
+                    return new RoomResponse
+                    {
+                        Success = false,
+                        ErrorCode = ErrorCode.CANNOT_DELETE_ROOM_WITH_BOOKINGS,
+                        Message = "Cannot delete room because it has existing bookings."
+                    };
+                }
+
+                var deleteSuccess = await _roomRepository.Delete(roomId);
+
+                if (deleteSuccess)
+                {
+                    return new RoomResponse
+                    {
+                        Success = true,
+                        Message = "The room was successfully deleted."
+                    };
+                }
 
                 return new RoomResponse
                 {
-                    Success = true,
-                    Message = "Sala deletada com sucesso."
+                    Success = false,
+                    Message = "Error occurred while trying to delete the room."
                 };
             }
             catch (Exception ex)
@@ -105,10 +127,11 @@ namespace Application.Room
                 return new RoomResponse
                 {
                     Success = false,
-                    Message = "Erro ao deletar sala: " + ex.Message
+                    Message = "Error deleting a room: " + ex.Message
                 };
             }
         }
+
 
 
         public async Task<RoomResponse> GetRoom(int roomId)
@@ -150,8 +173,78 @@ namespace Application.Room
             return responseList;
         }
 
+        public async Task<RoomResponse> UpdateRoom(int roomId, UpdateRoomRequest request)
+        {
+            try
+            {
+                var room = await _roomRepository.Get(roomId);
+                if (room == null)
+                {
+                    return new RoomResponse
+                    {
+                        Success = false,
+                        Message = "Room not found."
+                    };
+                }
 
-       
+                room.Name = request.RoomData.Name ?? room.Name;
+                room.Price = request.RoomData.PriceValue != 0
+                    ? new Price { Value = request.RoomData.PriceValue, Currency = (AcceptedCurrencies)request.RoomData.CurrencyCode }
+                    : room.Price;
+                room.InMaintenance = request.RoomData.InMaintenance;
+
+                room.Validate();
+
+                await _roomRepository.Update(room);
+
+                return new RoomResponse
+                {
+                    RoomData = RoomDto.MapToDto(room),
+                    Success = true,
+                    Message = "Room updated successfully!"
+                };
+            }
+            catch (InvalidRoomNameException)
+            {
+                return new RoomResponse
+                {
+                    Success = false,
+                    ErrorCode = ErrorCode.MISSING_REQUIRED_INFORMATION,
+                    Message = "Missing name information"
+                };
+            }
+            catch (InvalidRoomPriceException)
+            {
+                return new RoomResponse
+                {
+                    Success = false,
+                    ErrorCode = ErrorCode.INVALID_PRICE,
+                    Message = "The passed price cannot be <= 0"
+                };
+            }
+            catch (InvalidRoomLevelException)
+            {
+                return new RoomResponse
+                {
+                    Success = false,
+                    ErrorCode = ErrorCode.MISSING_REQUIRED_INFORMATION,
+                    Message = "Missing level information"
+                };
+            }
+            catch (Exception ex)
+            {
+                return new RoomResponse
+                {
+                    Success = false,
+                    Message = "Error updating room: " + ex.Message
+                };
+            }
+        }
+
+
+
+
+
 
     }
 }
